feat: Refactor project structure and add quote filtering by author
- Moved routes to modular structure under `routers/` folder.
- Added filtering functionality to `/quotes` endpoint using query parameters.
- Simplified `main.py` by delegating route prefixes and tags to individual routers.
- Enhanced maintainability with centralized router aggregation in `routers/__init__.py`.
- Improved project scalability and readability for future feature additions.
